Add support for "meta issuers".
These are separate from the fixed OIDC issuers, and they represent templates for *classes* of OIDC endpoints that we want to support, e.g. for EKS:
```
https://oidc.eks.*.amazonaws.com/id/*
```

The `*` character here will be used to match `[a-zA-Z0-9_-]+`, so no host or path delimiting characters are allowed to prevent attacks like:

```
https://oidc.eks.mattmoor.io/pwned.amazonaws.com/id/does-not-matter
```

We do NOT maintain OIDC verifiers for all of the possible endpoints, but we do keep an LRU cache to avoid the expensive discovery process on each request.

Related: https://github.com/sigstore/fulcio/issues/212
Signed-off-by: Matt Moore <mattomata@gmail.com>